Move check_debugger_output to the debugger module
